pyproject.toml
[tool.poetry]
name = "github-actions-utils"
version = "0.1.0"
description = "Debabrata Udemy Github Actions Utils"
authors = ["Debabrata Patnaik <debabrata_patnaik@live.com>"]
readme = "README.md"
package-mode = false
license = "Apache License Version 2.0"

[tool.poetry.dependencies]
python = "^3.13"
strawberry-graphql = {extras = ["debug-server", "fastapi"], version = "^0.252.0"}
confluent-kafka = "2.6.1"
pydantic = "^2.10.1"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
colorama = "^0.4.6"
hydra-core = "^1.3.2"
mongojet = "^0.2.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pylint = "^3.3.1"
testcontainers = "^4.8.2"
redpanda = "^0.6.0"
coverage = "^7.6.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
indent-string = "  "
output-format = "colorized"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
